requirements.txt
flask
pytesseract
Pillow
nltk
opencv-python
langdetect
